添加第1项为路径名字

+++ src/lib.rs
@@ -80,7 +80,7 @@ fn ensure_all_window_dark_mode() -> Result<usize> {
 }
 
 #[defun]
-fn popup_shell_menu(path: String, x: usize, y: usize) -> Result<usize> {
+fn popup_shell_menu(path: String, x: i32, y: i32) -> Result<usize> {
     if let Err(e) = crate::shellmenu::pop_shell_menu(path, x, y) {
         let es = format!("popup_shell_menu error: {}", e);
         let esw = to_wstring(&es);
